Move sorting logic to the API route file Introduce a types file to organize prisma types
